Backend-rag/requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
fastapi
uvicorn
pypdf
python-docx
python-dotenv
sentence-transformers
langchain-groq
python-multipart
langchain-huggingface
pydantic
rank-bm25
ragas
pytest
httpx